* master:
  feat(metric-stats): Report cardinality to metric stats (#3360)
  release: 0.8.56
  fix(perfscore): Adds span op tag to perf score totals (#3326)
  ref(profiles): Return retention_days as part of the Kafka message (#3362)
  ref(filter): Add GTmetrix to the list of web crawlers (#3363)
  fix: Fix kafka topic default (#3350)
  ref(normalization): Remove duplicated normalization (#3355)
  feat(feedback): Emit outcomes for user feedback events (#3026)
  feat(cardinality): Implement cardinality reporting (#3342)

+++ relay-cardinality/src/redis/state.rs
@@ -13,9 +13,10 @@ use crate::{
 ///
 /// Also tracks amount of cache hits and misses as well as amount of
 /// items accepted and rejected. These metrics are reported via statsd on drop.
+#[derive(Debug)]
 pub struct LimitState<'a> {
     /// The limit of the quota.
-    pub limit: u64,
+    pub limit: u32,
 
     /// Scoping of the quota.
     partial_scope: PartialQuotaScoping,
